MILP.Core

Description

This module contains the core ADT for defining MILP problems.

Types

data Model

A model consists of an objective function and a list of constraints.

Constructors

Model

Fields

mObj :: ObjF

mConstrs :: [Constraint]

type VarName = String

type Coeff = Float

type Var = (Coeff, VarName)

Variables are represented as the pair of a coefficient with a variable name.

data Term

A term is either a literal value or a variable. Lists of terms ([Term]) are considered to be summations of all the terms in the list.

Constructors

TLit Float

TVar Var

data CType

A constraint type can either be equality or an inequality.

Constructors

Eql

Geq

Leq

data Constraint

Constraints can represent an equality or inequality between two lists of terms or the constraint of a list of variables to binary values or to integer values.

Constructors

C CType [Term] [Term]

BinC [Var]

IntC [Var]

data ObjDir

Direction of the objective function is either minimize or maximize.

Constructors

ObjMin

ObjMax

data ObjF

An objective function consists of a direction and a list of terms.

Constructors

ObjF ObjDir [Term]

Creating Models

var :: Var -> Term

Create a Term from a Var.

var1 :: String -> Term

Create a Term from a variable name with a coefficient of 1.

varNeg1 :: String -> Term

Create a Term from a variable name with a coefficient of -1.

constant :: Float -> Term

Create a Term from a literal.

constantNeg :: Float -> Term

Concatenation of constant and negate.

geq :: [Term] -> [Term] -> Constraint

Create a constraint where the sum of the first list of terms is greater than or equal to the second list of terms.

leq :: [Term] -> [Term] -> Constraint

Create a constraint where the sum of the first list of terms is less than or equal to the second list of terms.

eql :: [Term] -> [Term] -> Constraint

Create a constraint where the sum of the first list of terms is equal to the second list of terms.

Writing Model

showModel :: Model -> String

Given a model creates a String in the format of an LPSolve file.

writeLP :: FilePath -> Model -> IO ()

Given a model, writes the result of showModel to the given path.

preprocess :: Model -> Model

Applies a few optimisations to the given model by removing redundant constraints and variables.