Recipe.Helper

Description

This module contains various helper functions for use with recipes.

General

ppList :: Show a => [a] -> IO ()

Prints a list with every element on a new line.

deps :: IxAction -> Recipe -> [IxAction]

Get the dependencies of a given action in the given recipe.

reverseR :: Recipe -> Recipe

Reverse all the edges in the given recipe graph.

Lifting function to work on recipes

liftR :: (Graph IxAction -> a) -> Recipe -> a

Lift a polymorphic function on graphs of actions to a function on recipes.

liftR' :: (Graph IxAction -> Graph IxAction) -> Recipe -> Recipe

Lift a function on graphs of actions to recipes.

liftR2' :: (Graph IxAction -> Graph IxAction -> Graph IxAction) -> Recipe -> Recipe -> Recipe

Lift a function over two graphs of actions to a function over two recipes.