Recipe.Kitchen

Description

This module defines the environment in which recipes are scheduled.

Types

data Station

Represents a station (such as an oven) within the environment.

Constructors

Station

Fields

stName :: String

Unique name

stFunc :: Action -> [Action] -> Maybe Time

Given an action and its dependencies returns the time take for the station to perform that action. Returns Nothing if the station cannot perform the given action.

stCapacity :: Int

Number of concurrent actions the station can perform.

data Env

An environment is simply a wrapper around a list of station in the environment.

Constructors

Env [Station]

Scheduling Helper Functions

fromCapacity :: Env -> Env

Splits a station down into one substation for each concurrent action it is capable of performing.

validStations :: IxAction -> Recipe -> Env -> [(String, Time)]

Given an action from a given recipe and an environment, produces a list of station names paired with the time it would take each station to perform the recipe.

validStations' :: IxAction -> Recipe -> Env -> [(Station, Time)]

Same as validStations but pairs the time with the station, not just the station name.

isValidSt :: Action -> [Action] -> Station -> Bool

Given an action and its dependencies returns True if the given station can perform the action.

intersectStations :: [(String, Time)] -> [(String, Time)] -> [(String, Time, Time)]

Creates the intersection of two lists of stations produced by validStations. The intersection only considers the station name hence both times are returned as they may be different.

infeasible :: Recipe -> Env -> Maybe [IxAction]

Returns a list of actions from the given recipe that cannot be performed by any station in the given environment.

maxEnd :: Recipe -> Env -> Time

Maximum end time of the recipe if every action was performed sequentially using the slowest station for each action.

maxDur :: Recipe -> Env -> Time

Maximum duration of any single action in the given recipe.