Recipe.Recipe

Description

This module defines the combinators used to create recipes.

Types

type IxAction = (Int, Action)

An action paired with a unique label.

newtype Recipe

A recipe is a wrapper around a graph of actions.

Constructors

R (Graph IxAction)

data Action

Constructors

GetIngredient String

Heat

HeatAt Temp

Wait

Mix

Conditional Condition Action

Transaction Action

Measure Measurement

Using [String] Action

Separate String

With

type Recipe' = (Recipe, IxAction)

A recipe paired with the last action to be added.

type STRecipe = State Int Recipe'

A recipe paired with the last action to be added with state keeping track of the unqiue label counter for actions.

newtype Temp

Temp is a wrapper around an Int representing degrees centigrade.

Constructors

Temp Int

newtype Time

Time is a wrapper around an Int representing seconds.

Constructors

Time Int

data Condition

Conditions represent an event which you perform an action "until".

Constructors

CondTime Time
Until the given time has elapsed.

CondTemp Temp
Until the temperature has been reached.

Condition `AND` Condition
Conjunction of two conditions.

Condition `OR` Condition
Disjunction of two conditions.

data Measurement

Represents a measurement of something.

Constructors

Count Int
Number of something e.g. 1 apple.

Grams Int
Number of grams.

Milliletres Int
Number of milliletres.

Combinators

mkRecipe :: STRecipe -> Recipe

Evaluate the state starting at 0 and generate a recipe.

ingredient :: String -> STRecipe

An ingredient with a given name.

heat :: STRecipe -> STRecipe

Heat dependencies, used in conjunction with conditional.

heatAt :: Temp -> STRecipe -> STRecipe

Heat dependencies at a given temperature, used in conjunction with conditional.

wait :: STRecipe -> STRecipe

Do nothing, used in conjunction with conditional.

mix :: STRecipe -> STRecipe -> STRecipe

Mix dependencies together.

conditional :: Condition -> STRecipe -> STRecipe

Wrap an action with a termination condition.

transaction :: STRecipe -> STRecipe

Wrapping an action in a transaction means that all dependencies must be completed around the same time and that the wrapped action must be started within a small time of the dependencies being completed.

measure :: Measurement -> STRecipe -> STRecipe

Measure the dependencies.

using :: [String] -> STRecipe -> STRecipe

Perform the given action using a station with one of the given names.

separate :: String -> String -> STRecipe -> State Int (Recipe', Recipe')

Used when a dependency produces multiple outputs to label which output is being referred to.

with :: STRecipe -> STRecipe -> STRecipe

Serve dependencies with each other.

Conditional Helper Functions

forTime :: Time -> STRecipe -> STRecipe

Wrap previously added action in condition of time.

toTemp :: Temp -> STRecipe -> STRecipe

Wrap previously added action in condition of temperature.

heatTo :: Temp -> STRecipe -> STRecipe

Concatenation of heat and toTemp.

heatFor :: Time -> STRecipe -> STRecipe

Concatenation of heat and forTime.

heatAtFor :: Temp -> Time -> STRecipe -> STRecipe

Concatenation of heatAt and forTime.

waitFor :: Time -> STRecipe -> STRecipe

Concatenation of wait and forTime.

Time Helper Functions

minutes :: Int -> Time

minutes 5 = Time 300

hours :: Int -> Time

hours 2 = Time 7200

Measurement Helper Functions

grams :: Int -> STRecipe -> STRecipe

grams 100 = measure (Grams 100)

ml :: Int -> STRecipe -> STRecipe

ml 200 = measure (Milliletres 200)

count :: Int -> STRecipe -> STRecipe

count 5 = measure (Count 5)

getMeasure :: Measurement -> Int

Returns the magnitude of a Measurement.

Multiple Output Helper Functions

discardFirst :: State Int (Recipe', Recipe') -> STRecipe

Discard the first output produced by the previously added action and keep the second.

discardSecond :: State Int (Recipe', Recipe') -> STRecipe

Discard the second output produced by the previously added action and keep the first.