Recipe.Scheduler.Heuristic

Description

This module uses a heuristic based approach to schedule recipes.

Types

type ActionRule = [IxAction] -> Recipe -> Env -> IxAction

Selects an action from the given list. Is provided with the full recipe and the environment in which the recipe is being scheduled.

type StationRule = [(String, Time)] -> IxAction -> Recipe -> Schedule -> (String, Time)

Selects from a list of station names paired with the time that the station requires to perform the given action. Is also given the full recipe and the current schedule.

type Schedule = [(String, [Task])]

A list of station names paired with their list of tasks.

data Task

A station is either idle for a given time or performing a given action for a given time.

Constructors

Idle Float

Active IxAction Float

Scheduling Recipes

ruleSchedule :: ActionRule -> StationRule -> Recipe -> Env -> Schedule

Schedule the given recipe in the given environment using the provided rules.

mkEmptySch :: Env -> Schedule

Given an environment, create a schedule in which all stations have no tasks.

scheduleLength :: Schedule -> Float

Length of a schedule (end time of final task).

Rules / Heuristics

shortestFirst :: ActionRule

Schedule the shortest action first.

longestFirst :: ActionRule

Schedule the longest action first.

leastFlexible :: ActionRule

Schedule the least flexible action first. That is the action which has the smallest number of valid stations.

fastestFirst :: StationRule

Choose the station which can perform the action the fastest.

leastIdleReq :: StationRule

Choose the station for which the least amount of idle time would need to be added.

leastUsed :: StationRule

Choose the station which has the fewest actions assigned to it.